+++ plugin/buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -1,9 +1,9 @@
 object PluginVersions {
-    val kotlin = "1.2.30"
-    val ktlintPlugin = "3.1.0"
+    val kotlin = "1.2.41"
+    val ktlintPlugin = "3.3.0"
     val gradlePublishPlugin = "0.9.10"
-    val androidPlugin = "3.0.1"
-    val kotlinNativePlugin = "0.6.1"
+    val androidPlugin = "3.1.1"
+    val kotlinNativePlugin = "0.7"
     val semver = "1.0.0"
-    val gradleWrapper = "4.6"
+    val gradleWrapper = "4.7"
 }