+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased [3.4.0-SNAPSHOT]
+### Added
+### Changed
+  - Update Kotlin to 1.2.41 version
+  - Update Gradle wrapper to 4.7 version
+### Fixed
+
 ## [3.3.0] - 2018-4-24
 ### Added
  - Check for spaces in output path for KtLint versions earlier